+++ src/styles/theme.ts
@@ -1,4 +1,6 @@
-import { lighten, darken, transparentize } from "polished";
+import lighten from "polished/lib/color/lighten";
+import darken from "polished/lib/color/darken";
+import transparentize from "polished/lib/color/transparentize";
 
 import { defaultColours } from "./DefaultColours";